regulate type coercion

+++ service/connect_manager/src/multimodal_input_connect_stub.cpp
@@ -39,9 +39,9 @@ int32_t MultimodalInputConnectStub::OnRemoteRequest(
     }
 
     switch (code) {
-        case static_cast<uint32_t>(IMultimodalInputConnect::ALLOC_SOCKET_FD):
+        case IMultimodalInputConnect::ALLOC_SOCKET_FD:
             return StubHandleAllocSocketFd(data, reply);
-        case static_cast<uint32_t>(IMultimodalInputConnect::SET_EVENT_POINTER_FILTER):
+        case IMultimodalInputConnect::SET_EVENT_POINTER_FILTER:
             return StubAddInputEventFilter(data, reply);
         default:
             MMI_LOGE("unknown code:%{public}u, go switch defaut", code);
